refactor(types): move module-scoped type declarations
Scoped type declarations are now moved into their respective modules
(e.g., `modules/artworks/types.ts`).

Global type files have been renamed to follow the same convention:
- types/auth.d.ts → types/auth.ts
- types/env.d.ts        (unchanged, remains valid)
- types/express.d.ts    (unchanged, remains valid)
- types/APIError.d.ts   (deleted, obsolete)

This has been documented in docs/type_file_conventions.md
