+++ src/modules/artworks/model.ts
@@ -1,9 +1,9 @@
 // vim: foldlevel=1:foldnestmax=2
 import debugFactory from 'debug';
 
-import client from '../../interfaces/db/pgClient.js';
-import APIError from '../../infrastructure/shared/APIError.js';
-import type { Artwork } from '@/types/artworks.d.js';
+import client from '@/interfaces/db/pgClient.js';
+import APIError from '@/infrastructure/shared/APIError.js';
+import type { Artwork } from './types.js';
 
 const debug = debugFactory('datamapper');
