+++ src/modules/artworks/types.ts
