--- src/types/APIError.d.ts
@@ -1,7 +1,0 @@
-export interface ApiError {
-  error: Error,
-  message: string,
-  code: number,
-  date: Date,
-  stack: string,
-}
